knex migrations to create tables
